refactor : product price 변수 타입 Double에서 int로 조정 가격이 int 타입을 넘지 않기때문에 조정

+++ src/main/java/com/sparta/goods/domain/product/dto/ProductRequest.java
@@ -13,13 +13,13 @@ import lombok.Setter;
 public class ProductRequest {
 
     private String productName;
-    private Double price;
+    private int price;
     private Integer quantity;
     private String introduction;
     private GoodsCategory category;
 
     @Builder
-    public ProductRequest(String productName, Double price, Integer quantity,
+    public ProductRequest(String productName, int price, Integer quantity,
                           String introduction, GoodsCategory category) {
 
         this.productName = productName;
