+++ src/main/java/com/sparta/goods/domain/product/entity/Product.java
@@ -22,11 +22,11 @@ public class Product {
 
     @Column
     @Min(100)
-    private Double price;
+    private int price;
 
     @Column
     @Min(0)
-    private Integer quantity;
+    private int quantity;
 
     @Column
     @NotNull
@@ -38,7 +38,7 @@ public class Product {
 
 
     @Builder
-    public Product(String productName, Double price, Integer quantity, String introduction, GoodsCategory category) {
+    public Product(String productName, int price, Integer quantity, String introduction, GoodsCategory category) {
         this.productName = productName;
         this.price = price;
         this.quantity = quantity;
